requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
python-dotenv==1.0.0
pydantic==2.3.0
python-multipart==0.0.6
redis==4.6.0
aioredis==2.0.1
qrcode==7.4.2
pillow==10.0.0
